requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.26.1
celery==3.1.23
coverage==3.7.1
dj-database-url==0.2.2
Django==1.6.2
django-annoying==0.8.1
django-bootstrap-form==3.2
django-bootstrap3==5.1.1
django-braces==1.4.0
-e git://github.com/jezdez/django-configurations@5ece1070448d3e026b81422691c60207f5699c0c#egg=django_configurations
django-debug-toolbar==1.0.1
django-dotenv==1.2
django-extensions==1.3.3
django-filer==0.9.9
django-fsm==2.5.0
django-getenv==1.1
django-mptt==0.6.1
django-polymorphic==0.6.1
django-reversion==1.8.0
django-secure==1.0
django-storages==1.1.8
django-tables2==0.15.0
django-widget-tweaks==1.4.1
djrill==1.3.0
easy-thumbnails==2.2
Fabric==1.7.0
future==0.14.3
gunicorn==18.0
icalendar==3.8.4
jsmin==2.1.4
jsonfield==0.9.20
kombu==3.0.35
lxml==3.1.0
newrelic==2.16.0.12
opbeat==1.2
paramiko==1.11.0
Pillow==2.7.0
psycopg2==2.5.2
pycrypto==2.6
python-docx==0.8.6
docxtpl==0.2.4
python-dateutil==2.4.0
pytz==2016.4
redis==2.10.3
requests==2.6.0
six==1.5.2
South==0.8.4
sqlparse==0.1.11
Unidecode==0.4.17
Werkzeug==0.8.3
